+++ src/app/domain/artwork/artwork.port.ts
@@ -25,6 +25,7 @@ export interface ArtworkPort extends Artwork {
     tokenId: string,
     lang: string,
     body: string,
+    title: string,
     token: string
   ): Observable<ArtCritic | null>;
 }